split source types into their own files in separate package
